pyproject.toml
[tool.poetry]
name = "turbo-text-transformer"
version = "0.1.6"
description = ""
authors = ["fergus <fergusfettes@gmail.com>"]
homepage = "https://github.com/fergusfettes/turbo-text-transformer"
readme = "README.md"
packages = [{include = "ttt"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<3.12"
pyyaml = "^6.0"
openai = "^0.27.0"
click = "^8.1.3"
colored = "^1.4.4"
tiktoken = "^0.3.0"
turbo-text-transformer-prompts = "^0.1.10"
langchain = {git = "ssh://git@github.com/fergusfettes/langchain", rev = "0.0.119"}
gpt_index = {git = "ssh://git@github.com/fergusfettes/llama_index", rev = "0.5.13"}
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.11"
ipython = "^8.10.0"
memory-profiler = "^0.61.0"

[tool.poetry.scripts]
ttt = "ttt.__main__:main"

[project]
name = "turbo-text-transformer"
version = "0.1.6"
dynamic = ["dependencies"]

[tool.setuptools.dynamic]
dependencies = {file = ["requirements.txt"]}

[project.scripts]
ttt = "ttt.__main__:main"

[tool.setuptools]
packages = ["ttt"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
